Remove unsetting of NetCoreTargetingPackRoot to allow package pruning (#70)
Contributes to #61 

Package pruning in .NET 10 requires `NetCoreTargetingPackRoot` to be
set. Previously, we unset that property to isolate the build environment
from machine configuration.

To resolve the issue, we're removing our unsetting of this property so
that package pruning can work as expected. The more officially supported
way to instead enforce a consistent SDK version is via global.json.
However, enforcing that is a bigger change that may break some
workflows.

As a result, we're going to ship this change in two pieces: this piece,
which is non-breaking and unblocks .NET 10 SDK adoption. The global.json
enforcement will come later in #69.